feat: Add "smart" configuration screen
I added another screen for setting the document type instead of the <Modal> component. This is due to usage of react-native-picker-select library and how visually pickers look on Android and iOS.

I tried to make it smart, i.e. 2nd picker is disabled until you make a choice in the 1st one. Also, options of the 2nd picker depend on the choice in the 1st one. Additionally, the button for the next transition is disabled until you make all the choices.

In order to check whether the parameters were passed properly, I added the caption "You selected "document type" which was issued in "document country"" caption on the ResultsScreen.

The code is understandable, but probably I still should try my best to make it cleaner and more readable.